[ALPS04743585] ccci: Invalid type in argument to printf format specifier
Force conversion to int.

MTK-Commit-Id: 585633d5043f238abf9d9cc273ec1cb19fab3a97

CR-Id: ALPS04743585
Feature: Modem Interface Driver

+++ drivers/misc/mediatek/eccci/port/port_proxy.c
@@ -379,7 +379,7 @@ long port_dev_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 						SMEM_USER_MD_WIFI_PROXY);
 		sub_smem->size &= ~(PAGE_SIZE - 1);
 		CCCI_NORMAL_LOG(port->md_id, TAG, "wifi smem size =%lx(%d)\n",
-			(unsigned long)sub_smem->size, PAGE_SIZE);
+			(unsigned long)sub_smem->size, (int)PAGE_SIZE);
 		ret = put_user((unsigned int)sub_smem->size,
 				(unsigned int __user *)arg);
 		break;
